# server/openapi.yaml
openapi: 3.1.0
info:
  title: GitHub Vote LLM API
  version: 1.0.0
  description: API for the GitHub Vote LLM
  contact:
    name: Didrik Olofsson
    email: didrik@didrik.dev
    url: https://didrik.dev
servers:
  - url: http://localhost:8080
    description: Local development
paths:
...
  /v1/api/github/webhook:
    post:
      summary: GitHub webhook
      description: Receives GitHub webhook events. Validates request using X-Hub-Signature-256.
      operationId: handleGithubWebhook
      tags:
        - Webhooks
      security:
        - hubSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: GitHub webhook event payload (varies by event type)
      responses:
        "200":
          description: Webhook processed successfully
        "401":
          description: Invalid or missing signature
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /v1/api/repos/{owner}/{repo}/config:
    get:
      summary: Get repo config
      description: Get configuration for a specific repository.
      operationId: getRepoConfig
      tags:
        - Repos
      security:
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/Owner"
        - $ref: "#/components/parameters/Repo"
      responses:
        "200":
          description: Repo configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Repo config not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          $ref: "#/components/responses/ApiKeyNotConfigured"
    delete:
      summary: Delete repo config
      description: Delete configuration for a repository.
      operationId: deleteRepoConfig
      tags:
        - Repos
      security:
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/Owner"
        - $ref: "#/components/parameters/Repo"
      responses:
        "204":
          description: Config deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ApiKeyNotConfigured"
    put:
      summary: Update repo config
      description: Update configuration for a repository. All fields are optional.
      operationId: updateRepoConfig
      tags:
        - Repos
      security:
        - apiKey: []
      parameters:
        - $ref: "#/components/parameters/Owner"
        - $ref: "#/components/parameters/Repo"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRepoConfigRequest"
      responses:
        "200":
          description: Config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepoConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ApiKeyNotConfigured"
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: API key for authenticating requests
    hubSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
      description: HMAC-SHA256 signature of the webhook payload
  parameters:
...
    Owner:
      name: owner
      in: path
      required: true
      schema:
        type: string
      description: Repository owner (e.g. GitHub org or username)
    Repo:
      name: repo
      in: path
      required: true
      schema:
        type: string
      description: Repository name
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Error message
...
    RepoConfig:
      type: object
      required: [id, owner, repo, label_approved, label_in_progress, label_done, label_failed, label_feature_request, vote_threshold, timeout_minutes, max_budget_usd, is_board_public, created_at, updated_at]
      properties:
        id:
          type: integer
          format: int64
        owner:
          type: string
        repo:
          type: string
        label_approved:
          type: string
        label_in_progress:
          type: string
        label_done:
          type: string
        label_failed:
          type: string
        label_feature_request:
          type: string
        vote_threshold:
          type: integer
          format: int32
        timeout_minutes:
          type: integer
          format: int32
        max_budget_usd:
          type: number
          format: double
        is_board_public:
          type: boolean
          description: Whether the community board is publicly accessible
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UpdateRepoConfigRequest:
      type: object
      description: All fields are optional. Only provided fields will be updated.
      properties:
        label_approved:
          type: string
        label_in_progress:
          type: string
        label_done:
          type: string
        label_failed:
          type: string
        label_feature_request:
          type: string
        vote_threshold:
          type: integer
          format: int32
        timeout_minutes:
          type: integer
          format: int32
        max_budget_usd:
          type: number
          format: double
        anthropic_api_key:
          type: string
          description: API key for Anthropic (stored securely, not returned in responses)
        is_board_public:
          type: boolean
          description: Whether the community board is publicly accessible
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: unauthorized
    ApiKeyNotConfigured:
      description: API key is not configured on the server
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: api key not configured